ref(ts) Convert organization teams to typescript
Remove some redundant properties that were unused and move away from
mixins to HoCs and function components.
